Update TaskSyncRecoverLoots.java
The way that this class worked before, the For loop would look through every single tile entity in the chunk, and if one of them was a hopper, it would assign it to the "hopper" variable. Only THEN would it check if that hopper was a Chunk Hopper, and if it wasn't, the entity's death/item drop would be completely discarded. As a result, if there was a chunk hopper and a normal hopper in the same chunk, there was a chance that the "hopper" variable would be assigned to the normal hopper but then invalidated, and the chunk hopper would not receive the drops. With this simple change, the check would be made WITHIN the For loop to make sure that the hopper was a Chunk Hopper rather than a normal hopper.